Add view for the job posting page

+++ SearchJob/Dtos/Job/CreateJobRequestDto.cs
@@ -9,9 +9,9 @@
 
         public string Location { get; set; }
 
-        public int? SalaryFrom { get; set; }
+        public decimal? SalaryFrom { get; set; }
 
-        public int? SalaryTo { get; set; }
+        public decimal? SalaryTo { get; set; }
 
         public string Url { get; set; }
