+++ SearchJob/Dtos/Job/JobDto.cs
@@ -11,9 +11,9 @@ namespace SearchJob.Dtos.Job
 
         public string Location { get; set; } 
 
-        public int? SalaryFrom { get; set; }
+        public decimal? SalaryFrom { get; set; }
 
-        public int? SalaryTo { get; set; }
+        public decimal? SalaryTo { get; set; }
 
         public string Url { get; set; }
